Created new hook for storing compare state
