+++ src/components/Scheduler/index.tsx
@@ -10,7 +10,7 @@ import {
 } from '..';
 import { OverlayCrnsContext, OverlayCrnsContextValue } from '../../contexts';
 import { DESKTOP_BREAKPOINT } from '../../constants';
-import useLocalStorageState from 'use-local-storage-state';
+import useCompareStateFromStorage from '../../data/hooks/useCompareStateFromStorage';
 import useScreenWidth from '../../hooks/useScreenWidth';
 
 /**
@@ -39,49 +39,8 @@ export default function Scheduler(): React.ReactElement {
   //   }
   // );
 
-  const [compare, setCompare] = useLocalStorageState<boolean>(
-    'compare-panel-state-compareValue',
-    {
-      defaultValue: false,
-      storageSync: false,
-    }
-  );
-  const [pinnedSchedules, setPinnedSchedules] = useLocalStorageState<string[]>(
-    'compare-panel-state-pinnedSchedules',
-    {
-      defaultValue: [],
-      storageSync: false,
-    }
-  );
-  const [pinSelf, setPinSelf] = useLocalStorageState<boolean>(
-    'compare-panel-state-pinSelfValue',
-    {
-      defaultValue: true,
-      storageSync: false,
-    }
-  );
-
-  console.log(compare);
-
-  const handleCompareSchedules = useCallback(
-    (
-      newCompare?: boolean,
-      newPinnedSchedules?: string[],
-      newPinSelf?: boolean
-    ) => {
-      if (newCompare !== undefined) {
-        console.log(newCompare);
-        setCompare(newCompare);
-      }
-      if (newPinnedSchedules !== undefined) {
-        setPinnedSchedules(newPinnedSchedules);
-      }
-      if (newPinSelf !== undefined) {
-        setPinSelf(newPinSelf);
-      }
-    },
-    []
-  );
+  const { compare, pinned, pinSelf, handleCompareSchedules } =
+    useCompareStateFromStorage();
 
   return (
     <>
@@ -108,7 +67,7 @@ export default function Scheduler(): React.ReactElement {
                 className="calendar"
                 overlayCrns={overlayCrns}
                 compare={compare}
-                pinnedFriendSchedules={pinnedSchedules}
+                pinnedFriendSchedules={pinned}
                 pinSelf={!compare || pinSelf}
               />
             </div>
@@ -116,7 +75,7 @@ export default function Scheduler(): React.ReactElement {
           {(!mobile || tabIndex === 3) && (
             <ComparisonPanel
               handleCompareSchedules={handleCompareSchedules}
-              pinnedSchedules={pinnedSchedules}
+              pinnedSchedules={pinned}
               pinSelf={pinSelf}
               compare={compare}
             />
